Add form for downtime notification
