+++ ffmd/gluon-config-mode-notify-2016.2/Makefile
@@ -1,6 +1,6 @@
 include $(TOPDIR)/rules.mk
 
-PKG_NAME:=gluon-config-mode-contact-info
+PKG_NAME:=gluon-config-mode-notify
 PKG_VERSION:=1
 
 PKG_BUILD_DIR := $(BUILD_DIR)/$(PKG_NAME)
@@ -10,16 +10,15 @@ include $(GLUONDIR)/include/package.mk
 PKG_CONFIG_DEPENDS += $(GLUON_I18N_CONFIG)
 
 
-define Package/gluon-config-mode-contact-info
+define Package/gluon-config-mode-notify
   SECTION:=gluon
   CATEGORY:=Gluon
-  TITLE:=Allows the user to provide contact information to be distributed in the mesh
+  TITLE:=Allows the user to select if they want to be notified about problems with this node
   DEPENDS:=gluon-config-mode-core-virtual +gluon-node-info
 endef
 
-define Package/gluon-config-mode-contact-info/description
-	Allows the user to provide contact information to be distributed in the mesh.
-	Can be made obligatory in site.conf
+define Package/gluon-config-mode-notify/description
+	Allows the user to select if they want to be notified about problems with this node.
 endef
 
 define Build/Prepare
@@ -30,18 +29,18 @@ define Build/Configure
 endef
 
 define Build/Compile
-	$(call GluonBuildI18N,gluon-config-mode-contact-info,i18n)
+	$(call GluonBuildI18N,gluon-config-mode-notify,i18n)
 	$(call GluonSrcDiet,./luasrc,$(PKG_BUILD_DIR)/luadest/)
 endef
 
-define Package/gluon-config-mode-contact-info/install
+define Package/gluon-config-mode-notify/install
 	$(CP) $(PKG_BUILD_DIR)/luadest/* $(1)/
-	$(call GluonInstallI18N,gluon-config-mode-contact-info,$(1))
+	$(call GluonInstallI18N,gluon-config-mode-notify,$(1))
 endef
 
-define Package/gluon-config-mode-contact-info/postinst
+define Package/gluon-config-mode-notify/postinst
 #!/bin/sh
 $(call GluonCheckSite,check_site.lua)
 endef
 
-$(eval $(call BuildPackage,gluon-config-mode-contact-info))
+$(eval $(call BuildPackage,gluon-config-mode-notify))
