+++ ffmd/gluon-config-mode-notify-2016.2/check_site.lua
@@ -1,3 +1,3 @@
-if need_table('config_mode', nil, false) and need_table('config_mode.owner', nil, false) then
+if need_table('config_mode', nil, false) and need_table('owner', nil, false) then
   need_boolean('config_mode.owner.obligatory', false)
 end
